Title: 닷 투더 닷, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241022/닷 투더 닷/dot-to-the-dot.cpp
@@ -55,7 +55,7 @@ int main() {
         }
     }
 
-    cout << dist[n];
+    cout << (int)dist[n];
 
     return 0;
 }
